moved mv command to common dso interface
